+++ fidget/src/mesh/octree.rs
@@ -1363,7 +1363,7 @@ impl LeafHermiteData {
 mod test {
     use super::*;
     use crate::{
-        context::bound::{self, BoundContext, BoundNode},
+        context::Tree,
         eval::{EzShape, MathShape},
         mesh::types::{Edge, X, Y, Z},
         shape::Bounds,
@@ -1391,12 +1391,8 @@ mod test {
         },
     };
 
-    fn sphere(
-        ctx: &BoundContext,
-        center: [f32; 3],
-        radius: f32,
-    ) -> bound::BoundNode {
-        let (x, y, z) = ctx.axes();
+    fn sphere(center: [f32; 3], radius: f32) -> Tree {
+        let (x, y, z) = Tree::axes();
         ((x - center[0]).square()
             + (y - center[1]).square()
             + (z - center[2]).square())
@@ -1404,13 +1400,8 @@ mod test {
             - radius
     }
 
-    fn cube(
-        ctx: &BoundContext,
-        bx: [f32; 2],
-        by: [f32; 2],
-        bz: [f32; 2],
-    ) -> BoundNode {
-        let (x, y, z) = ctx.axes();
+    fn cube(bx: [f32; 2], by: [f32; 2], bz: [f32; 2]) -> Tree {
+        let (x, y, z) = Tree::axes();
         let x_bounds = (bx[0] - x.clone()).max(x - bx[1]);
         let y_bounds = (by[0] - y.clone()).max(y - by[1]);
         let z_bounds = (bz[0] - z.clone()).max(z - bz[1]);
@@ -1420,12 +1411,10 @@ mod test {
     #[test]
     fn test_cube_edge() {
         const EPSILON: f32 = 1e-3;
-        let ctx = BoundContext::new();
         let f = 2.0;
-        let cube = cube(&ctx, [-f, f], [-f, 0.3], [-f, 0.6]);
+        let shape = VmShape::from_tree(cube([-f, f], [-f, 0.3], [-f, 0.6]));
         // This should be a cube with a single edge running through the root
         // node of the octree, with an edge vertex at [0, 0.3, 0.6]
-        let shape: VmShape = cube.convert();
         let octree = Octree::build(&shape, DEPTH0_SINGLE_THREAD);
         assert_eq!(octree.verts.len(), 5);
         let v = octree.verts[0].pos;
@@ -1437,19 +1426,18 @@ mod test {
     }
 
     fn cone(
-        ctx: &BoundContext,
         corner: nalgebra::Vector3<f32>,
         tip: nalgebra::Vector3<f32>,
         radius: f32,
-    ) -> BoundNode {
+    ) -> Tree {
         let dir = tip - corner;
         let length = dir.norm();
         let dir = dir.normalize();
 
-        let corner = corner.map(|v| ctx.constant(v as f64));
-        let dir = dir.map(|v| ctx.constant(v as f64));
+        let corner = corner.map(|v| Tree::constant(v as f64));
+        let dir = dir.map(|v| Tree::constant(v as f64));
 
-        let (x, y, z) = ctx.axes();
+        let (x, y, z) = Tree::axes();
         let point = nalgebra::Vector3::new(x, y, z);
         let offset = point.clone() - corner.clone();
 
@@ -1473,9 +1461,7 @@ mod test {
 
     #[test]
     fn test_mesh_basic() {
-        let ctx = BoundContext::new();
-        let shape = sphere(&ctx, [0.0; 3], 0.2);
-        let shape: VmShape = shape.convert();
+        let shape = VmShape::from_tree(sphere([0.0; 3], 0.2));
 
         // If we only build a depth-0 octree, then it's a leaf without any
         // vertices (since all the corners are empty)
@@ -1518,10 +1504,8 @@ mod test {
 
     #[test]
     fn test_sphere_verts() {
-        let ctx = BoundContext::new();
-        let shape = sphere(&ctx, [0.0; 3], 0.2);
+        let shape = VmShape::from_tree(sphere([0.0; 3], 0.2));
 
-        let shape: VmShape = shape.convert();
         let octree = Octree::build(&shape, DEPTH1_SINGLE_THREAD);
         let sphere_mesh = octree.walk_dual(DEPTH1_SINGLE_THREAD);
 
@@ -1556,9 +1540,7 @@ mod test {
 
     #[test]
     fn test_sphere_manifold() {
-        let ctx = BoundContext::new();
-        let shape = sphere(&ctx, [0.0; 3], 0.85);
-        let shape: VmShape = shape.convert();
+        let shape = VmShape::from_tree(sphere([0.0; 3], 0.85));
 
         for threads in [0, 8] {
             let settings = Settings {
@@ -1587,10 +1569,9 @@ mod test {
 
     #[test]
     fn test_cube_verts() {
-        let ctx = BoundContext::new();
-        let shape = cube(&ctx, [-0.1, 0.6], [-0.2, 0.75], [-0.3, 0.4]);
+        let shape =
+            VmShape::from_tree(cube([-0.1, 0.6], [-0.2, 0.75], [-0.3, 0.4]));
 
-        let shape: VmShape = shape.convert();
         let octree = Octree::build(&shape, DEPTH1_SINGLE_THREAD);
         let mesh = octree.walk_dual(DEPTH1_SINGLE_THREAD);
         const EPSILON: f32 = 2.0 / u16::MAX as f32;
@@ -1634,13 +1615,12 @@ mod test {
     #[test]
     fn test_plane_center() {
         const EPSILON: f32 = 1e-3;
-        let ctx = BoundContext::new();
         for dx in [0.0, 0.25, -0.25, 2.0, -2.0] {
             for dy in [0.0, 0.25, -0.25, 2.0, -2.0] {
                 for offset in [0.0, -0.2, 0.2] {
-                    let (x, y, z) = ctx.axes();
+                    let (x, y, z) = Tree::axes();
                     let f = x * dx + y * dy + z + offset;
-                    let shape: VmShape = f.convert();
+                    let shape = VmShape::from_tree(f);
                     let octree = Octree::build(&shape, DEPTH0_SINGLE_THREAD);
 
                     assert_eq!(octree.cells.len(), 8);
@@ -1675,10 +1655,8 @@ mod test {
             nalgebra::Vector3::new(0.2, 0.3, 0.4),
             nalgebra::Vector3::new(1.2, 1.3, 1.4),
         ] {
-            let ctx = BoundContext::new();
             let corner = nalgebra::Vector3::new(-1.0, -1.0, -1.0);
-            let shape = cone(&ctx, corner, tip, 0.1);
-            let shape: VmShape = shape.convert();
+            let shape = VmShape::from_tree(cone(corner, tip, 0.1));
 
             let mut eval = VmShape::new_point_eval();
             let tape = shape.ez_point_tape();
@@ -1704,12 +1682,10 @@ mod test {
     fn test_mesh_manifold() {
         for threads in [0, 8] {
             for i in 0..256 {
-                let ctx = BoundContext::new();
                 let mut shape = vec![];
                 for j in Corner::iter() {
                     if i & (1 << j.index()) != 0 {
                         shape.push(sphere(
-                            &ctx,
                             [
                                 if j & X { 0.5 } else { 0.0 },
                                 if j & Y { 0.5 } else { 0.0 },
@@ -1724,7 +1700,7 @@ mod test {
 
                 // Now, we have our shape, which is 0-8 spheres placed at the
                 // corners of the cell spanning [0, 0.25]
-                let shape: VmShape = shape.convert();
+                let shape = VmShape::from_tree(shape);
                 let settings = Settings {
                     min_depth: 2,
                     max_depth: 2,
@@ -1751,33 +1727,28 @@ mod test {
 
     #[test]
     fn test_collapsible() {
-        let ctx = BoundContext::new();
-
-        fn builder(
-            shape: BoundNode,
-            settings: Settings,
-        ) -> OctreeBuilder<VmShape> {
-            let shape: VmShape = shape.convert();
+        fn builder(shape: Tree, settings: Settings) -> OctreeBuilder<VmShape> {
+            let shape = VmShape::from_tree(shape);
             let eval = Arc::new(EvalGroup::new(shape));
             let mut out = OctreeBuilder::new();
             out.recurse(&eval, CellIndex::default(), settings);
             out
         }
 
-        let shape = sphere(&ctx, [0.0; 3], 0.1);
+        let shape = sphere([0.0; 3], 0.1);
         let octree = builder(shape, DEPTH1_SINGLE_THREAD);
         assert!(!octree.collapsible(8));
 
-        let shape = sphere(&ctx, [-1.0; 3], 0.1);
+        let shape = sphere([-1.0; 3], 0.1);
         let octree = builder(shape, DEPTH1_SINGLE_THREAD);
         assert!(octree.collapsible(8));
 
-        let shape = sphere(&ctx, [-1.0, 0.0, 1.0], 0.1);
+        let shape = sphere([-1.0, 0.0, 1.0], 0.1);
         let octree = builder(shape, DEPTH1_SINGLE_THREAD);
         assert!(!octree.collapsible(8));
 
-        let a = sphere(&ctx, [-1.0; 3], 0.1);
-        let b = sphere(&ctx, [1.0; 3], 0.1);
+        let a = sphere([-1.0; 3], 0.1);
+        let b = sphere([1.0; 3], 0.1);
         let shape = a.min(b);
         let octree = builder(shape, DEPTH1_SINGLE_THREAD);
         assert!(!octree.collapsible(8));
@@ -1786,9 +1757,7 @@ mod test {
     #[test]
     fn test_empty_collapse() {
         // Make a very smol sphere that won't be sampled
-        let ctx = BoundContext::new();
-        let shape = sphere(&ctx, [0.1; 3], 0.05);
-        let tape: VmShape = shape.convert();
+        let shape = VmShape::from_tree(sphere([0.1; 3], 0.05));
         for threads in [0, 4] {
             let settings = Settings {
                 min_depth: 1,
@@ -1796,7 +1765,7 @@ mod test {
                 threads,
                 ..Default::default()
             };
-            let octree = Octree::build(&tape, settings);
+            let octree = Octree::build(&shape, settings);
             assert_eq!(
                 octree.cells[0],
                 Cell::Empty.into(),
@@ -1906,10 +1875,8 @@ mod test {
 
     #[test]
     fn test_qef_near_planar() {
-        let ctx = BoundContext::new();
-        let shape = sphere(&ctx, [0.0; 3], 0.75);
+        let shape = VmShape::from_tree(sphere([0.0; 3], 0.75));
 
-        let shape: VmShape = shape.convert();
         let settings = Settings {
             min_depth: 4,
             max_depth: 4,
@@ -1926,10 +1893,8 @@ mod test {
 
     #[test]
     fn test_octree_bounds() {
-        let ctx = BoundContext::new();
-        let shape = sphere(&ctx, [1.0; 3], 0.25);
+        let shape = VmShape::from_tree(sphere([1.0; 3], 0.25));
 
-        let shape: VmShape = shape.convert();
         let center = Vector3::new(1.0, 1.0, 1.0);
         let settings = Settings {
             min_depth: 4,